# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wedding Invitation API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type GuestSalutationType =
  (typeof GuestSalutationType)[keyof typeof GuestSalutationType];

export const GuestSalutationType = {
  Дорогой: "Дорогой",
  Дорогая: "Дорогая",
  Дорогие: "Дорогие",
} as const;

export type GuestRsvpStatus =
  (typeof GuestRsvpStatus)[keyof typeof GuestRsvpStatus];

export const GuestRsvpStatus = {
  pending: "pending",
  attending: "attending",
  not_attending: "not_attending",
} as const;

export interface Guest {
  id: number;
  firstName: string;
  lastName: string;
  salutationType: GuestSalutationType;
  guestsCount: number;
  slug: string;
  invitationOpened: boolean;
  gameCompleted: boolean;
  rsvpStatus: GuestRsvpStatus;
  /** @nullable */
  rsvpComment: string | null;
  /** @nullable */
  tableId: number | null;
  /** @nullable */
  seatNumber: number | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateGuestBodySalutationType =
  (typeof CreateGuestBodySalutationType)[keyof typeof CreateGuestBodySalutationType];

export const CreateGuestBodySalutationType = {
  Дорогой: "Дорогой",
  Дорогая: "Дорогая",
  Дорогие: "Дорогие",
} as const;

export interface CreateGuestBody {
  firstName: string;
  lastName: string;
  salutationType: CreateGuestBodySalutationType;
  guestsCount?: number;
  /** @nullable */
  slug?: string | null;
}

export type UpdateGuestBodySalutationType =
  (typeof UpdateGuestBodySalutationType)[keyof typeof UpdateGuestBodySalutationType];

export const UpdateGuestBodySalutationType = {
  Дорогой: "Дорогой",
  Дорогая: "Дорогая",
  Дорогие: "Дорогие",
} as const;

export type UpdateGuestBodyRsvpStatus =
  (typeof UpdateGuestBodyRsvpStatus)[keyof typeof UpdateGuestBodyRsvpStatus];

export const UpdateGuestBodyRsvpStatus = {
  pending: "pending",
  attending: "attending",
  not_attending: "not_attending",
} as const;

export interface UpdateGuestBody {
  firstName?: string;
  lastName?: string;
  salutationType?: UpdateGuestBodySalutationType;
  guestsCount?: number;
  slug?: string;
  rsvpStatus?: UpdateGuestBodyRsvpStatus;
  /** @nullable */
  rsvpComment?: string | null;
  /** @nullable */
  tableId?: number | null;
  /** @nullable */
  seatNumber?: number | null;
}

export type RsvpBodyRsvpStatus =
  (typeof RsvpBodyRsvpStatus)[keyof typeof RsvpBodyRsvpStatus];

export const RsvpBodyRsvpStatus = {
  attending: "attending",
  not_attending: "not_attending",
} as const;

export interface RsvpBody {
  rsvpStatus: RsvpBodyRsvpStatus;
  /** @nullable */
  rsvpComment?: string | null;
}

export interface Settings {
  id: number;
  weddingTitle: string;
  brideName: string;
  groomName: string;
  weddingDate: string;
  weddingTime: string;
  venueName: string;
  venueAddress: string;
  venueLat: number;
  venueLng: number;
  /** @nullable */
  mapLink: string | null;
  invitationText: string;
  /** @nullable */
  dressCode: string | null;
  /** @nullable */
  contacts: string | null;
  adminPassword: string;
  gameEnabled: boolean;
  countdownEnabled: boolean;
  updatedAt: string;
}

export interface UpdateSettingsBody {
  weddingTitle?: string;
  brideName?: string;
  groomName?: string;
  weddingDate?: string;
  weddingTime?: string;
  venueName?: string;
  venueAddress?: string;
  venueLat?: number;
  venueLng?: number;
  /** @nullable */
  mapLink?: string | null;
  invitationText?: string;
  /** @nullable */
  dressCode?: string | null;
  /** @nullable */
  contacts?: string | null;
  adminPassword?: string;
  gameEnabled?: boolean;
  countdownEnabled?: boolean;
}

export interface GuestStats {
  total: number;
  attending: number;
  notAttending: number;
  pending: number;
  gameCompleted: number;
  invitationOpened: number;
  totalPersons: number;
  attendingPersons: number;
  notAttendingPersons: number;
  pendingPersons: number;
}

export interface WeddingTable {
  id: number;
  name: string;
  seatsCount: number;
  sortOrder: number;
  /** @nullable */
  note: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTableBody {
  name: string;
  seatsCount?: number;
  sortOrder?: number;
  /** @nullable */
  note?: string | null;
}

export interface UpdateTableBody {
  name?: string;
  seatsCount?: number;
  sortOrder?: number;
  /** @nullable */
  note?: string | null;
}

export interface ActivityLog {
  id: number;
  guestId: number;
  guestName: string;
  eventType: string;
  /** @nullable */
  payload: string | null;
  createdAt: string;
}

export interface AdminLoginBody {
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  token: string;
}