# lib/api-zod/src/generated/types/activityLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wedding Invitation API
 * OpenAPI spec version: 0.2.0
 */

export interface ActivityLog {
  id: number;
  guestId: number;
  guestName: string;
  eventType: string;
  /** @nullable */
  payload: string | null;
  createdAt: Date;
}