# lib/api-zod/src/generated/types/createGuestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wedding Invitation API
 * OpenAPI spec version: 0.2.0
 */
import type { CreateGuestBodySalutationType } from "./createGuestBodySalutationType";

export interface CreateGuestBody {
  firstName: string;
  lastName: string;
  salutationType: CreateGuestBodySalutationType;
  guestsCount?: number;
  /** @nullable */
  slug?: string | null;
}