# lib/api-zod/src/generated/types/guest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wedding Invitation API
 * OpenAPI spec version: 0.2.0
 */
import type { GuestRsvpStatus } from "./guestRsvpStatus";
import type { GuestSalutationType } from "./guestSalutationType";

export interface Guest {
  id: number;
  firstName: string;
  lastName: string;
  salutationType: GuestSalutationType;
  guestsCount: number;
  slug: string;
  invitationOpened: boolean;
  gameCompleted: boolean;
  rsvpStatus: GuestRsvpStatus;
  /** @nullable */
  rsvpComment: string | null;
  /** @nullable */
  tableId: number | null;
  /** @nullable */
  seatNumber: number | null;
  createdAt: Date;
  updatedAt: Date;
}