# lib/api-zod/src/generated/types/guestStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wedding Invitation API
 * OpenAPI spec version: 0.2.0
 */

export interface GuestStats {
  total: number;
  attending: number;
  notAttending: number;
  pending: number;
  gameCompleted: number;
  invitationOpened: number;
  totalPersons: number;
  attendingPersons: number;
  notAttendingPersons: number;
  pendingPersons: number;
}