# lib/api-zod/src/generated/types/rsvpBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wedding Invitation API
 * OpenAPI spec version: 0.2.0
 */
import type { RsvpBodyRsvpStatus } from "./rsvpBodyRsvpStatus";

export interface RsvpBody {
  rsvpStatus: RsvpBodyRsvpStatus;
  /** @nullable */
  rsvpComment?: string | null;
}