# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wedding Invitation API
 * OpenAPI spec version: 0.2.0
 */

export interface Settings {
  id: number;
  weddingTitle: string;
  brideName: string;
  groomName: string;
  weddingDate: string;
  weddingTime: string;
  venueName: string;
  venueAddress: string;
  venueLat: number;
  venueLng: number;
  /** @nullable */
  mapLink: string | null;
  invitationText: string;
  /** @nullable */
  dressCode: string | null;
  /** @nullable */
  contacts: string | null;
  adminPassword: string;
  gameEnabled: boolean;
  countdownEnabled: boolean;
  updatedAt: Date;
}