# lib/api-zod/src/generated/types/weddingTable.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wedding Invitation API
 * OpenAPI spec version: 0.2.0
 */

export interface WeddingTable {
  id: number;
  name: string;
  seatsCount: number;
  sortOrder: number;
  /** @nullable */
  note: string | null;
  createdAt: Date;
  updatedAt: Date;
}